typecheck: Delete old test file
